requirements.txt
aniso8601==1.2.1
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
coverage==4.4.1
coveralls==1.2.0
cryptography==2.0.3
docopt==0.6.2
enum34==1.1.6
Flask==0.12.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
funcsigs==1.0.2
geopy==1.11.0
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
json-logging-py==0.2
MarkupSafe==1.0
mock==2.0.0
pbr==3.1.1
py==1.4.34
pycparser==2.18
pyfcm==1.4.3
PyJWT==1.4.2
pymongo==3.5.1
pyOpenSSL==17.2.0
pytest==3.2.1
pytest-cov==2.5.1
python-dateutil==2.6.1
python-firebase==1.2
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
